Removed dupe unit tests
