+++ Tests/UnitTestsParallelizable/View/Adornment/ToScreenTests.cs
@@ -1,0 +1,14 @@
+﻿using Xunit.Abstractions;
+
+namespace Terminal.Gui.ViewTests;
+
+/// <summary>
+/// Test the <see cref="Adornment.FrameToScreen"/> and <see cref="View.ViewportToScreen"/> methods.
+/// DOES NOT TEST View.xxxToScreen methods. Those are in ./View/Layout/ToScreenTests.cs
+/// </summary>
+/// <param name="output"></param>
+public class AdornmentToScreenTests (ITestOutputHelper output)
+{
+    private readonly ITestOutputHelper _output = output;
+
+}
